// backend/src/types/activity.types.ts
import { Document, Types } from "mongoose";

export enum ActivityAction {
  PROJECT_CREATED = "PROJECT_CREATED",
  PROJECT_UPDATED = "PROJECT_UPDATED",
  PROJECT_DELETED = "PROJECT_DELETED",

  TASK_CREATED = "TASK_CREATED",
  TASK_UPDATED = "TASK_UPDATED",
  TASK_ASSIGNED = "TASK_ASSIGNED",
  TASK_COMPLETED = "TASK_COMPLETED",
  TASK_DELETED = "TASK_DELETED",

  COMMENT_ADDED = "COMMENT_ADDED",

  ISSUE_CREATED = "ISSUE_CREATED",
  ISSUE_UPDATED = "ISSUE_UPDATED",
  ISSUE_CLOSED = "ISSUE_CLOSED",
  ISSUE_DELETED = "ISSUE_DELETED",

  REPOSITORY_CREATED = "REPOSITORY_CREATED",
  REPOSITORY_UPDATED = "REPOSITORY_UPDATED",
  REPOSITORY_DELETED = "REPOSITORY_DELETED",

  TEAM_CREATED = "TEAM_CREATED",
  MEMBER_ADDED = "MEMBER_ADDED",
  MEMBER_REMOVED = "MEMBER_REMOVED",

  MEETING_CREATED = "MEETING_CREATED",
  MEETING_UPDATED = "MEETING_UPDATED",
  MEETING_CANCELLED = "MEETING_CANCELLED",

  SYSTEM_ACTIVITY = "SYSTEM_ACTIVITY",
}

export enum ActivityEntityType {
  PROJECT = "PROJECT",
  TASK = "TASK",
  COMMENT = "COMMENT",
  ISSUE = "ISSUE",
  REPOSITORY = "REPOSITORY",
  TEAM = "TEAM",
  MEETING = "MEETING",
}

export interface IActivity extends Document {
  user: Types.ObjectId;
  project?: Types.ObjectId;

  action: ActivityAction;
  description: string;

  entityType?: ActivityEntityType;
  entityId?: Types.ObjectId;

  createdAt: Date;
  updatedAt: Date;
}